refactor(estree/tokens): break up into multiple files (#19850)
Pure refactor. Just break up implementation of tokens serialization into multiple files.
